src/requirements.txt
asn1crypto==0.24.0
certifi==2018.1.18
cffi==1.11.5
chardet==3.0.4
click==6.7
cryptography==2.2
enum34==1.1.6
Flask==0.12.2
httplib2==0.10.3
idna==2.6
ipaddress==1.0.19
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
pycparser==2.18
pyOpenSSL==17.5.0
requests==2.18.4
six==1.11.0
urllib3==1.22
Werkzeug==0.14.1
